# cmd/gomodel/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "High-performance AI gateway routing requests to multiple LLM providers (OpenAI, Anthropic, Gemini, Groq, xAI, Oracle, Ollama). Drop-in OpenAI-compatible API.",
        "title": "GOModel API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/v1/batches": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "batch"
                ],
                "summary": "List batches",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Pagination cursor",
                        "name": "after",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Maximum items to return (1-100, default 20)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/core.BatchListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/core.GatewayError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/core.GatewayError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/core.GatewayError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/core.GatewayError"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "batch"
                ],
                "summary": "Create a native provider batch",
                "parameters": [
                    {
                        "description": "Batch request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/core.BatchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/core.BatchResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/core.GatewayError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/core.GatewayError"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/core.GatewayError"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    },
    "definitions": {
...
        "core.BatchError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "core.BatchListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/core.BatchResponse"
                    }
                },
                "first_id": {
                    "type": "string"
                },
                "has_more": {
                    "type": "boolean"
                },
                "last_id": {
                    "type": "string"
                },
                "object": {
                    "type": "string"
                }
            }
        },
        "core.BatchRequest": {
            "type": "object",
            "properties": {
                "completion_window": {
                    "type": "string"
                },
                "endpoint": {
                    "type": "string"
                },
                "input_file_id": {
                    "type": "string"
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "requests": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/core.BatchRequestItem"
                    }
                }
            }
        },
        "core.BatchRequestCounts": {
            "type": "object",
            "properties": {
                "completed": {
                    "type": "integer"
                },
                "failed": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "core.BatchRequestItem": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "object"
                },
                "custom_id": {
                    "type": "string"
                },
                "method": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "core.BatchResponse": {
            "type": "object",
            "properties": {
                "cancelled_at": {
                    "type": "integer"
                },
                "cancelling_at": {
                    "type": "integer"
                },
                "completed_at": {
                    "type": "integer"
                },
                "completion_window": {
                    "type": "string"
                },
                "created_at": {
                    "type": "integer"
                },
                "endpoint": {
                    "type": "string"
                },
                "failed_at": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "in_progress_at": {
                    "type": "integer"
                },
                "input_file_id": {
                    "type": "string"
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "object": {
                    "type": "string"
                },
                "provider": {
                    "type": "string"
                },
                "provider_batch_id": {
                    "type": "string"
                },
                "request_counts": {
                    "$ref": "#/definitions/core.BatchRequestCounts"
                },
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/core.BatchResultItem"
                    }
                },
                "status": {
                    "type": "string"
                },
                "usage": {
                    "description": "Gateway extension: optional usage/result snapshots persisted by the gateway.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/core.BatchUsageSummary"
                        }
                    ]
                }
            }
        },
        "core.BatchResultItem": {
            "type": "object",
            "properties": {
                "custom_id": {
                    "type": "string"
                },
                "error": {
                    "$ref": "#/definitions/core.BatchError"
                },
                "index": {
                    "type": "integer"
                },
                "model": {
                    "type": "string"
                },
                "provider": {
                    "type": "string"
                },
                "response": {},
                "status_code": {
                    "type": "integer"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "core.BatchUsageSummary": {
            "type": "object",
            "properties": {
                "input_cost": {
                    "type": "number"
                },
                "input_tokens": {
                    "type": "integer"
                },
                "output_cost": {
                    "type": "number"
                },
                "output_tokens": {
                    "type": "integer"
                },
                "total_cost": {
                    "type": "number"
                },
                "total_tokens": {
                    "type": "integer"
                }
            }
        },
...
        "core.ErrorType": {
            "type": "string",
            "enum": [
                "provider_error",
                "rate_limit_error",
                "invalid_request_error",
                "authentication_error",
                "not_found_error"
            ],
            "x-enum-varnames": [
                "ErrorTypeProvider",
                "ErrorTypeRateLimit",
                "ErrorTypeInvalidRequest",
                "ErrorTypeAuthentication",
                "ErrorTypeNotFound"
            ]
        },
...
        "core.GatewayError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "x-nullable": true
                },
                "message": {
                    "type": "string"
                },
                "param": {
                    "type": "string",
                    "x-nullable": true
                },
                "provider": {
                    "type": "string"
                },
                "status_code": {
                    "type": "integer"
                },
                "type": {
                    "$ref": "#/definitions/core.ErrorType"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}